Feat : Merge 후 간단한 코드 리뷰.
